pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiocryptopay>=0.4.8",
    "aiogram>=3.25.0",
    "asyncpg>=0.31.0",
    "fpdf2>=2.8.5",
    "openpyxl>=3.1.5",
    "pyotp>=2.9.0",
]
